Split and rename files
