+++ src/dryad/format/endf/createBoundary.hpp
@@ -1,5 +1,5 @@
-#ifndef NJOY_DRYAD_FORMAT_ENDF_BOUNDARIES
-#define NJOY_DRYAD_FORMAT_ENDF_BOUNDARIES
+#ifndef NJOY_DRYAD_FORMAT_ENDF_CREATEBOUNDARY
+#define NJOY_DRYAD_FORMAT_ENDF_CREATEBOUNDARY
 
 // system includes
 #include <vector>
@@ -25,20 +25,6 @@ namespace endf {
       throw std::exception();
     }
     return static_cast< std::size_t >( boundary ) - 1;
-  }
-
-  /**
-   *  @brief Convert a range of ENDF boundary indices to normal boundary indeces
-   */
-  template < typename Range >
-  auto createBoundaries( const Range& boundaries )
-  -> std::enable_if_t< njoy::tools::std20::ranges::range< Range >,
-                       std::vector< std::size_t > > {
-
-    std::vector< std::size_t > converted( boundaries.size() );
-    std::transform( boundaries.begin(), boundaries.end(),
-                    converted.begin(), &createBoundary );
-    return converted;
   }
 
 } // endf namespace
